# refactor: improve performance when parsing header