--- a/src/node/src/lib.rs
+++ b/src/node/src/lib.rs
@@ -159,6 +159,7 @@ pub fn parse_voice(path_or_buf: Either<String, Buffer>) -> napi::Result<HashMap<
     huffman_lookup_table: &vec![],
     order_by_steamid: false,
     fallback_bytes: None,
+    parse_grenades: false
   };
   let mut parser = Parser::new(settings, parser::parse_demo::ParsingMode::Normal);
   let output = parse_demo(bytes, &mut parser)?;
@@ -194,6 +195,7 @@ pub fn list_game_events(path_or_buf: Either<String, Buffer>) -> napi::Result<Val
     huffman_lookup_table: &huf,
     order_by_steamid: false,
     fallback_bytes: None,
+    parse_grenades: false
   };
   let mut parser = Parser::new(settings, parser::parse_demo::ParsingMode::Normal);
   let output = parse_demo(bytes, &mut parser)?;
@@ -238,6 +240,7 @@ pub fn parse_grenades(
     huffman_lookup_table: &huf,
     order_by_steamid: false,
     fallback_bytes: None,
+    parse_grenades: false
   };
   let mut parser = Parser::new(settings, parser::parse_demo::ParsingMode::Normal);
   let output = parse_demo(bytes, &mut parser)?;
@@ -286,6 +289,7 @@ pub fn parse_header(path_or_buf: Either<String, Buffer>) -> napi::Result<Value> 
     huffman_lookup_table: &huf,
     order_by_steamid: false,
     fallback_bytes: None,
+    parse_grenades: false
   };
   let mut parser = Parser::new(settings, parser::parse_demo::ParsingMode::Normal);
   let output = parse_demo(bytes, &mut parser)?;
@@ -359,6 +363,7 @@ pub fn parse_event(
     huffman_lookup_table: &huf,
     order_by_steamid: false,
     fallback_bytes: game_event_list_bytes,
+    parse_grenades: false
   };
   let mut parser = Parser::new(settings, parser::parse_demo::ParsingMode::Normal);
   let output = parse_demo(bytes, &mut parser)?;
@@ -430,6 +435,7 @@ pub fn parse_events(
     huffman_lookup_table: &huf,
     order_by_steamid: false,
     fallback_bytes: game_event_list_bytes,
+    parse_grenades: false
   };
   let mut parser = Parser::new(settings, parser::parse_demo::ParsingMode::Normal);
   let output = parse_demo(bytes, &mut parser)?;
@@ -509,6 +515,7 @@ pub fn parse_ticks(
     huffman_lookup_table: &huf,
     order_by_steamid: order_by_steamid,
     fallback_bytes: None,
+    parse_grenades: false
   };
 
   let mut parser = Parser::new(settings, parser::parse_demo::ParsingMode::Normal);
@@ -583,6 +590,7 @@ pub fn parse_player_info(path_or_buf: Either<String, Buffer>) -> napi::Result<Va
     huffman_lookup_table: &huf,
     order_by_steamid: false,
     fallback_bytes: None,
+    parse_grenades: false
   };
   let mut parser = Parser::new(settings, parser::parse_demo::ParsingMode::Normal);
   let output = parse_demo(bytes, &mut parser)?;
@@ -614,6 +622,7 @@ pub fn parse_player_skins(path_or_buf: Either<String, Buffer>) -> napi::Result<V
     huffman_lookup_table: &huf,
     order_by_steamid: false,
     fallback_bytes: None,
+    parse_grenades: false
   };
   let mut parser = Parser::new(settings, parser::parse_demo::ParsingMode::Normal);
   let output = parse_demo(bytes, &mut parser)?;
@@ -644,6 +653,7 @@ pub fn list_updated_fields(path_or_buf: Either<String, Buffer>) -> napi::Result<
     huffman_lookup_table: &huf,
     order_by_steamid: false,
     fallback_bytes: None,
+    parse_grenades: false
   };
   let mut parser = Parser::new(settings, parser::parse_demo::ParsingMode::Normal);
   let output = parse_demo(bytes, &mut parser)?;