--- a/src/wasm/src/lib.rs
+++ b/src/wasm/src/lib.rs
@@ -1,4 +1,5 @@
 use parser::first_pass::parser_settings::rm_user_friendly_names;
+use parser::first_pass::parser_settings::FirstPassParser;
 use parser::first_pass::parser_settings::ParserInputs;
 use parser::parse_demo::Parser;
 use parser::parse_demo::ParsingMode::ForceSingleThreaded;
@@ -365,15 +366,10 @@ pub fn parseHeader(file: Vec<u8>) -> Result<JsValue, JsError> {
         fallback_bytes: None,
         parse_grenades: false
     };
-    let mut parser = Parser::new(settings, ForceSingleThreaded);
-    let output = match parser.parse_demo(&file) {
-        Ok(output) => output,
-        Err(e) => return Err(JsError::new(&format!("{}", e))),
-    };
+    let mut parser = FirstPassParser::new(&settings);
+    let output = parser.parse_header_only(&file).unwrap();
     let mut hm: HashMap<String, String> = HashMap::default();
-    if let Some(header) = output.header {
-        hm.extend(header);
-    }
+    hm.extend(output);
     match serde_wasm_bindgen::to_value(&hm) {
         Ok(s) => Ok(s),
         Err(e) => return Err(JsError::new(&format!("{}", e))),