Currently parser loops over entire file when parsing headers, which adds a lot of overhead.
Solution is to actually only parse header without reading entire file.

Speed improvements on 300mb demo:
In rust it goes from ~100ms to ~0.5ms
In node to ~4ms 
In wasm to ~50ms

Also this PR fixes builds for node and wasm, because in https://github.com/LaihoE/demoparser/pull/276 the bindings were not updated. **But I didn't expose new feature in node/wasm. It's only for build to not fail.**